Move HiveResult out of QueryExecution
